frysk-core/frysk/isa/corefiles/ChangeLog 2008-02-08  Andrew Cagney  <cagney@redhat.com>
	* IA32LinuxElfCorefile.java: Update to match Stat.ppid.`
	* PPC64LinuxElfCorefile.java: Ditto.
	* X8664LinuxElfCorefile.java: Ditto.
	* PPC32LinuxElfCorefile.java: Ditto.

frysk-core/frysk/proc/live/ChangeLog
2008-02-08  Andrew Cagney  <cagney@redhat.com>

	* LinuxPtraceTask.java: Use ProcssIdentifierFactory.create.
	* LinuxPtraceHost.java: Update to match Stat.ppid.

frysk-core/frysk/testbed/ChangeLog
2008-02-08  Andrew Cagney  <cagney@redhat.com>

	* OffspringType.java: Use ProcessIdentifierFactory.create(int).
	* TestLib.java: Update to match Stat.ppid.

frysk-sys/frysk/sys/ChangeLog
2008-02-08  Andrew Cagney  <cagney@redhat.com>

	* ProcessIdentifierFactory.java (create(int)): Make public.
	(createFIXME(int)): Delete.

frysk-sys/frysk/sys/proc/ChangeLog
2008-02-08  Andrew Cagney  <cagney@redhat.com>

	* Stat.java (scan(ProcessIdentifier)): New.
	(scan(ProcessIdentifier,ProcessIdentifier)): New.
	(tid, ppid): Change type to ProcessIdentifier.

frysk-sys/frysk/testbed/ChangeLog
2008-02-08  Andrew Cagney  <cagney@redhat.com>

	* TearDownProcess.java: Use ProcessIdentifierFactory.create(int).

+++ frysk-core/frysk/isa/corefiles/IA32LinuxElfCorefile.java
@@ -127,7 +127,7 @@ public class IA32LinuxElfCorefile extends LinuxElfCorefile {
 	prpsInfo.setPrGid(Status.getGID(pid));
 
 	prpsInfo.setPrPid(pid);
-	prpsInfo.setPrPpid(processStat.ppid);
+	prpsInfo.setPrPpid(processStat.ppid.intValue());
 	prpsInfo.setPrPgrp(processStat.pgrp);
 
 	prpsInfo.setPrSid(processStat.session);
@@ -174,7 +174,7 @@ public class IA32LinuxElfCorefile extends LinuxElfCorefile {
 
 	// Set initial prstatus data
 	prStatus.setPrPid(task.getTid());
-	prStatus.setPrPpid(processStat.ppid);
+	prStatus.setPrPpid(processStat.ppid.intValue());
 	prStatus.setPrPgrp(processStat.pgrp);
 	prStatus.setPrSid(processStat.session);
 	prStatus.setPrSigPending(processStat.signal);
